requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework==3.11.0
python-decouple
drf-yasg
